team_report_app/requirements.txt
Flask
Flask-SQLAlchemy
python-docx
reportlab
